added all code for issue
