+++ src/main/kotlin/ru/anarcom/octopus/service/UserServiceImpl.kt
@@ -81,6 +81,6 @@ class UserServiceImpl(
     override fun existsByEmail(email: String): Boolean =
         userRepository.existsByEmail(email)
 
-    override fun findUserByEmail(email: String): User? =
+    override fun findByEmail(email: String): User? =
         userRepository.findByEmail(email)
 }
